Working! YAYYYY
Butttttt

you need to run the program with path to the file, and then mv out.bin
to where hello_wayland project is, make sure hello_wayland is compiled
in just the right way (I have it stored under way cooler for now, will
move it to the src directory after this), and then run that program.

Oh, also the mouse makes it break because I didn't bother putting the
pointer image in the output of the .bin, because idc.

But hey it works! All I gotta do this is write this data into a wayland
surface, and we'll be good to go!