Add parsing for 'color' argument values in base 10
In the example configuration for way-cooler, the default background is
expressed as a lua numeric literal. This causes way-cooler-bg to take
a numeric literal rather than a hex string as input causing a behavior
where the background is set to a color other than what the user thinks
it should be from the configuration.

Change this behavior by changing the parsing of the color argument to
first check if it is a valid base 10 value, and then try to parse it as
base 16.